Revert "Display user's email on profile page"
This reverts commit 08a432997164e3522e01c2b9c03205b21698cb66.
