+++ app/package.json
@@ -20,8 +20,7 @@
     "body-parser": "^1.19.0",
     "ejs": "^3.0.1",
     "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "node-fetch": "^2.6.0"
+    "jsonwebtoken": "^8.5.1"
   },
   "devDependencies": {
     "eslint": "^6.6.0",
