+++ app/server/auth.js
@@ -1,10 +1,24 @@
 const jwt = require("jsonwebtoken");
 
-const { getCookie } = require("./util.js");
-
 const privateKey = process.env.secret || "secret";
 const TOKEN_NAME = "authToken";
 const JWT_ALGORITHM = "HS256";
+
+function getCookie(cookie, cookies) {
+  if (!cookies) {
+    return null;
+  }
+
+  // Search for the authentication token in the cookies
+  let value = null;
+  cookies.split(" ").forEach(c => {
+    const [cookieName, cookieValue] = c.split("=");
+    if (cookieName === cookie) {
+      value = cookieValue;
+    }
+  });
+  return value;
+}
 
 function authenticateRequest(req, res, next) {
   if (typeof req.headers.cookie !== "undefined") {
