+++ app/server/server.js
@@ -1,13 +1,8 @@
 const express = require("express");
 const path = require("path");
-const fetch = require("node-fetch");
 
 const app = express();
 const bodyParser = require("body-parser");
-
-const { getCookie } = require("./util.js");
-
-const API_ADDRESS = process.env.API_ADDRESS || "http://localhost:8080/graphql";
 
 const {
   authenticateRequest,
@@ -30,32 +25,8 @@ app.get("/login", authenticateLoggedOut, (req, res) => {
   res.render("pages/login");
 });
 
-app.get("/profile", authenticateRequest, async (req, res) => {
-  const query = `
-  {
-		me{
-			account{
-				email
-			}
-		}
-  }
-  `;
-  const cookies = req.headers.cookie;
-  const authToken = getCookie("authToken", cookies);
-  try {
-    const response = await fetch(API_ADDRESS, {
-      method: "POST",
-      headers: { "content-type": "application/json", authToken },
-      body: JSON.stringify({ query })
-    });
-    const json = await response.json();
-    const { email } = json.data.me.account;
-
-    res.render("pages/profile", { user: { email } });
-  } catch (error) {
-    console.log(error);
-    res.render("partials/error");
-  }
+app.get("/profile", authenticateRequest, (req, res) => {
+  res.render("pages/profile");
 });
 
 app.get("/register", authenticateLoggedOut, (req, res) => {
